Rename src/bartender/_ssh_utils.py -> src/bartender/ssh_utils.py
No need to mark it as private.
